[#2] update: split models from sources
